[rqd] Rework rssUpdate based on sessionid (#1736)
Partially revert changes from
https://github.com/AcademySoftwareFoundation/OpenCue/pull/1689 and
implement a different fix for the same problem based on sessionid.

The previous solution replaced a logic that traced lineage based on
sessionid with a parentid based solution. This new solution uses
sessionid but removes the session query loop from the main frame loop,
reducing the complexity of the logic and improving its performance.