[rqd] Ensure logdir is owned by job user (#1742)
Failing to do so causes problems if other frames from the same job were
executed on docker, which may lead to the log dir to be owned by root.

This change attempts to change the logdir ownership prior to
initializing `RqdLogger`. Handling this internally on the initialization
logic would be preferred, but is not possible as the initialization
needs to run as the job user and the chown logic needs to happen on
permissionHigh mode.

---------

Signed-off-by: Diego Tavares <dtavares@imageworks.com>
Co-authored-by: Ramon Figueiredo <ramon.fgrd@gmail.com>